Added QueueLike and a (currently hidden) state to Ctx
