+++ parsley-core/src/ghc/Parsley/Internal/Common.hs
@@ -10,13 +10,15 @@ Stability   : unstable
 module Parsley.Internal.Common (
     module Parsley.Internal.Common.Fresh,
     module Parsley.Internal.Common.Indexed,
-    module Parsley.Internal.Common.Queue,
+    module Parsley.Internal.Common.QueueLike, Queue, RewindQueue,
     module Parsley.Internal.Common.Utils,
     module Parsley.Internal.Common.Vec
   ) where
 
 import Parsley.Internal.Common.Fresh
 import Parsley.Internal.Common.Indexed
-import Parsley.Internal.Common.Queue
+import Parsley.Internal.Common.QueueLike
+import Parsley.Internal.Common.Queue (Queue)
+import Parsley.Internal.Common.RewindQueue (RewindQueue)
 import Parsley.Internal.Common.Utils
 import Parsley.Internal.Common.Vec
